// src/root/types.ts
export type RootState = {
  api: ApiState;
  apiToken: AuthState;
  application: ApplicationState;
  areaNote: AreaNoteState;
  areaSearch: AreaSearchState;
  auditLog: AuditLogState;
  batchrun: BatchRunState;
  billingPeriod: BillingPeriodState;
  collectionCourtDecision: CollectionCourtDecisionState;
  collectionLetter: CollectionLetterState;
  collectionNote: CollectionNoteState;
  comment: CommentState;
  contact: ContactState;
  contractFile: ContractFileState;
  createCollectionLetter: CreateCollectionLetterState;
  creditDecision: CreditDecisionState;
  district: DistrictState;
  form: Record<string, any>;
  index: IndexState;
  infillDevelopment: InfillDevelopmentState;
  infillDevelopmentAttachment: InfillDevelopmentAttachmentState;
  invoice: InvoiceState;
  invoiceNote: InvoiceNoteState;
  invoiceSet: InvoiceSetState;
  landUseContract: LandUseContractState;
  lease: LeaseState;
  leaseCreateCharge: LeaseCreateChargeState;
  leaseholdTransfer: LeaseholdTransferState;
  leaseType: LeaseTypeState;
  lessor: LessorState;
  penaltyInterest: PenaltyInterestState;
  plotSearch: PlotSearchState;
  plotApplications: PlotApplicationsState;
  previewInvoices: PreviewInvoicesState;
  rentBasis: RentBasisState;
  rentForPeriod: RentForPeriodState;
  sapInvoice: SapInvoicesState;
  serviceUnits: ServiceUnitsState;
  leaseStatisticReport: LeaseStatisticReportState;
  tradeRegister: TradeRegisterState;
  uiData: UiDataState;
  user: UserState;
  usersPermissions: UsersPermissionsState;
  vat: VatState;
  landUseInvoice: LandUseInvoicesState;
  landUseAgreementAttachment: LandUseAgreementAttachmentState;
};